Add files via upload
This code is focused in resolving linear equation systems using Gaussian elimination, similar to the previous code. However, I made it in C language to enhance my coding knowledge and be well acquainted with pointers, structures, dynamic allocation of memory and using matrices. There are some special conditions to guarantee the proper function of the code. Those demands are explicited in the description and in the READ.me. They are:  
- Adding the equations, there are specific ways to input them regarding the correct reading of the coefficients and correct attachment of the variables.
When using negative coefficients, do not separe the signal from the number; 
Evidentiate if the coefficient is equal to 1; 
Try to keep the same order of variables when entering the equations; 
In those cases where there are equations with some variables missing from others, evidentiate the presence of the variable by using '0' as coefficient. (You can put the variables together or separated from the coefficient).

An example of correct input is below: 

"Enter the equation 1 (use only one letter per variable): 1x + 8y + 0z = 10
Enter the equation 2 (use only one letter per variable): 0x -2y + 10z = -3"